Refactored, expanded view working with indicators, fixed left click position
